Water chemistry misbehaving -- closes #705
I had to fix the initializer for the Salt object. The amount_is_weight,
the amount and the addTo fields were all kind of cross wired.

The harder problem was the removeDuplicates method. I was unsure of the
side effects of removing the deleted and not displayed tests, so I
created a new method called `removeDuplicatesIgnoreDisplay` and used
that in the SaltTableModel.

Things seem to be working better now.