Tweaked FindBoost command in CMakeLists.txt
This commit repairs the configuration on Linux machines where we
have a recent CMake toolset (VERSION >= 3.29) where FindBoost.cmake
module is removed from standard installation.
Setting the policy CMP0167 helps ensure compatibility with older
versions of CMake as well.

Note : in this commit we also uniformise Boost retrievala across
platforms, not as COMPONENT but as a general dependency.

Signed-off-by: bebenlebricolo <benoittarrade@hotmail.fr>